requirements.txt
python-dotenv==1.2.1
openai==2.8.1
fastmcp==2.13.2